refactor(Section): mueve los CompGraficos a la carpeta de Section
