+++ src/Section/Section.jsx
@@ -1,8 +1,8 @@
 import React, { useContext, useState, useEffect } from 'react';
-import CompGraficos from '../CompGraficoLinea';
-import CompGraficoPie from '../CompGraficoPie';
-import CompGraficoBarra from '../CompGraficoBarra';
-import CompGraficoBarraHorizontal from '../CompGraficoBarraHorizontal';
+import CompGraficos from './CompGraficoLinea';
+import CompGraficoPie from './CompGraficoPie';
+import CompGraficoBarra from './CompGraficoBarra';
+import CompGraficoBarraHorizontal from './CompGraficoBarraHorizontal';
 import api from '../api';
 import AppContext from '../AppContext'
